[mesh][metrics] refactor engine metrics into observability metrics - move worker /metrics scraping and aggregation into observability metrics - move health and engine metrics public routes behind MetricsRouteFactory - keep the legacy core metrics aggregator entry point as a compatible re-export
