+++ atom/mesh/src/observability/metrics/mod.rs
@@ -1,6 +1,8 @@
 pub mod config;
+pub mod engine_metrics;
 pub mod mesh_metrics;
 pub mod recorder;
+pub mod routes;
 pub mod schema;
 
 #[path = "../metrics.rs"]
@@ -8,6 +10,7 @@ mod legacy;
 
 pub use config::PrometheusConfig;
 pub use recorder::{MeshMetrics, Metrics, StreamingMetricsParams};
+pub use routes::MetricsRouteFactory;
 pub use schema::{
     METRIC_INVENTORY, MetricKind, MetricSpec, MetricStatus, bool_to_static_str,
     labels as metrics_labels, method_to_static_str, normalize_path, status_code_to_cow,
