+++ atom/mesh/src/server.rs
@@ -14,25 +14,24 @@ use axum::{
     Json, Router,
 };
 use serde::Deserialize;
-use serde_json::{json, Value};
+use serde_json::Value;
 use tokio::{signal, spawn};
 use tracing::{debug, info, warn, Level};
 use wfaas::LoggingSubscriber;
 
 use crate::{
     app_context::AppContext,
-    config::{RouterConfig, RoutingMode},
+    config::RouterConfig,
     core::{
         job_queue::{JobQueue, JobQueueConfig},
         steps::{TokenizerConfigRequest, WorkflowEngines},
-        worker::WorkerType,
         worker_manager::WorkerManager,
         Job,
     },
     middleware::{self, QueuedRequest},
     observability::{
         logging::{self, LoggingConfig},
-        metrics::{self, PrometheusConfig},
+        metrics::{self, MetricsRouteFactory, PrometheusConfig},
     },
     protocols::{
         chat::ChatCompletionRequest,
@@ -76,63 +75,6 @@ async fn parse_reasoning(
 
 async fn sink_handler() -> Response {
     StatusCode::NOT_FOUND.into_response()
-}
-
-async fn liveness() -> Response {
-    (StatusCode::OK, "OK").into_response()
-}
-
-async fn readiness(State(state): State<Arc<AppState>>) -> Response {
-    let workers = state.context.worker_registry.get_all();
-    let healthy_workers: Vec<_> = workers.iter().filter(|w| w.is_healthy()).collect();
-
-    let is_ready = match &state.context.router_config.mode {
-        RoutingMode::PrefillDecode { .. } => {
-            let has_prefill = healthy_workers
-                .iter()
-                .any(|w| matches!(w.worker_type(), WorkerType::Prefill { .. }));
-            let has_decode = healthy_workers
-                .iter()
-                .any(|w| matches!(w.worker_type(), WorkerType::Decode));
-            has_prefill && has_decode
-        }
-        RoutingMode::Regular { .. } => !healthy_workers.is_empty(),
-    };
-
-    if is_ready {
-        (
-            StatusCode::OK,
-            Json(json!({
-                "status": "ready",
-                "healthy_workers": healthy_workers.len(),
-                "total_workers": workers.len()
-            })),
-        )
-            .into_response()
-    } else {
-        (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({
-                "status": "not ready",
-                "reason": "insufficient healthy workers"
-            })),
-        )
-            .into_response()
-    }
-}
-
-async fn health(_state: State<Arc<AppState>>) -> Response {
-    liveness().await
-}
-
-async fn health_generate(State(state): State<Arc<AppState>>, req: Request) -> Response {
-    state.router.health_generate(req).await
-}
-
-async fn engine_metrics(State(state): State<Arc<AppState>>) -> Response {
-    WorkerManager::get_engine_metrics(&state.context.worker_registry, &state.context.client)
-        .await
-        .into_response()
 }
 
 async fn get_server_info(State(state): State<Arc<AppState>>, req: Request) -> Response {
@@ -517,12 +459,8 @@ pub fn build_app(
             middleware::concurrency_limit_middleware,
         ));
 
-    let public_routes = Router::new()
-        .route("/liveness", get(liveness))
-        .route("/readiness", get(readiness))
-        .route("/health", get(health))
-        .route("/health_generate", get(health_generate))
-        .route("/engine_metrics", get(engine_metrics))
+    let public_routes = MetricsRouteFactory
+        .get(app_state.clone())
         .route("/v1/models", get(v1_models))
         .route("/get_model_info", get(get_model_info))
         .route("/get_server_info", get(get_server_info));
